number int for float

+++ js/init.js
@@ -1,5 +1,5 @@
-   var renderer,camera,obj;
-   onload=()=>{
+    var renderer,camera,obj;
+    onload=()=>{
         //================================BEGIN===================================
         var height=500; 
         var width=500;
@@ -17,31 +17,31 @@
         obj = new ObjectPredefinido();
         obj.draw(scene,renderer,camera);
         renderer.render(scene, camera);
-   };
-   Math.seno = function(g) {
-	return Math.sin(g * Math.PI / 180);
-   }
-   var aplicarMatriz=((matrizParaAplicar)=>{
+    };
+    Math.seno = function(g) {
+	    return Math.sin(g * Math.PI / 180);
+    }
+    var aplicarMatriz=((matrizParaAplicar)=>{
         scene = new THREE.Scene();
         // obj = new ObjectPredefinido(scene);
         new Eixos(scene);
         obj.aplicarMatriz(matrizParaAplicar);     
         obj.draw(scene,renderer,camera);
-   });
-   document.getElementById("aplicar").addEventListener("click",()=>{
+    });
+    document.getElementById("aplicar").addEventListener("click",()=>{
 		var matrizParaAplicar=[];
 		var puro = document.getElementById("entrada").value;
 		puro=puro.split("\n");
 		for(var l of puro){
 			var novo = l.split(',').map(function(item) {
-				return parseInt(item,10);
+				return parseFloat(item,10);
 			});
 			matrizParaAplicar.push(novo);
 		}
 		matrizParaAplicar.pop();
 		aplicarMatriz(matrizParaAplicar);
-   });
+    });
    
-   document.getElementById("resetar").addEventListener("click",()=>{
-		window.location="index.html";
-	});
+    document.getElementById("resetar").addEventListener("click",()=>{
+        window.location="index.html";
+    });
